Move topography tests to test_grid module and parameterize a few more tests
