+++ roms_tools/tests/test_setup/test_grid.py
@@ -3,8 +3,11 @@ import logging
 import textwrap
 from pathlib import Path
 
+import numpy as np
+import numpy.testing as npt
 import pytest
 import xarray as xr
+from scipy.ndimage import label
 
 from conftest import calculate_file_hash
 from roms_tools import Grid
@@ -14,6 +17,7 @@ from roms_tools.constants import (
     UPPER_BOUND_THETA_S,
 )
 from roms_tools.download import download_test_data
+from roms_tools.setup.topography import _compute_rfactor
 
 
 @pytest.fixture()
@@ -31,6 +35,86 @@ def clockwise_rotated_grid():
 
     grid = Grid(
         nx=1, ny=1, size_x=100, size_y=100, center_lon=-20, center_lat=0, rot=-20
+    )
+
+    return grid
+
+
+@pytest.fixture()
+def grid_that_straddles_dateline_with_shifted_global_etopo_data():
+
+    grid = Grid(
+        nx=5,
+        ny=5,
+        size_x=1000,
+        size_y=1000,
+        center_lon=0,
+        center_lat=0,
+        rot=20,
+        topography_source={
+            "name": "ETOPO5",
+            "path": download_test_data("etopo5_coarsened_and_shifted.nc"),
+        },
+    )
+
+    return grid
+
+
+@pytest.fixture()
+def grid_that_straddles_dateline_with_global_srtm15_data():
+
+    grid = Grid(
+        nx=5,
+        ny=5,
+        size_x=1000,
+        size_y=1000,
+        center_lon=0,
+        center_lat=0,
+        rot=20,
+        topography_source={
+            "name": "SRTM15",
+            "path": download_test_data("srtm15_coarsened.nc"),
+        },
+    )
+
+    return grid
+
+
+@pytest.fixture()
+def grid_that_straddles_180_degree_meridian_with_shifted_global_etopo_data():
+
+    grid = Grid(
+        nx=5,
+        ny=5,
+        size_x=1000,
+        size_y=1000,
+        center_lon=180,
+        center_lat=0,
+        rot=20,
+        topography_source={
+            "name": "ETOPO5",
+            "path": download_test_data("etopo5_coarsened_and_shifted.nc"),
+        },
+    )
+
+    return grid
+
+
+@pytest.fixture()
+def grid_that_straddles_180_degree_meridian_with_global_srtm15_data():
+
+    grid = Grid(
+        nx=5,
+        ny=5,
+        size_x=1000,
+        size_y=1000,
+        center_lon=180,
+        center_lat=0,
+        rot=20,
+        topography_source={
+            "name": "SRTM15",
+            "path": download_test_data("srtm15_coarsened.nc"),
+        },
     )
 
     return grid
@@ -80,6 +164,23 @@ def test_coords_relation(grid_fixture, request):
     assert_larger_equal_than_with_tolerance(grid.ds.lat_v.min(), grid.ds.lat_rho.min())
     assert_smaller_equal_than_with_tolerance(grid.ds.lon_v.max(), grid.ds.lon_rho.max())
     assert_smaller_equal_than_with_tolerance(grid.ds.lon_v.max(), grid.ds.lon_rho.max())
+
+
+@pytest.mark.parametrize(
+    "grid_fixture",
+    [
+        "grid_that_straddles_dateline",
+        "grid_that_straddles_180_degree_meridian",
+        "grid_that_straddles_dateline_with_shifted_global_etopo_data",
+        "grid_that_straddles_180_degree_meridian_with_shifted_global_etopo_data",
+        "grid_that_straddles_dateline_with_global_srtm15_data",
+        "grid_that_straddles_180_degree_meridian_with_global_srtm15_data",
+    ],
+)
+def test_successful_initialization_with_topography(grid_fixture, request):
+
+    grid = request.getfixturevalue(grid_fixture)
+    assert grid is not None
 
 
 def test_plot_save_methods(tmp_path):
@@ -220,22 +321,19 @@ def test_compatability_with_matlab_grid(tmp_path):
             (filepath.with_suffix(".nc")).unlink()
 
 
-def test_roundtrip_netcdf(tmp_path):
+@pytest.mark.parametrize(
+    "grid_fixture",
+    [
+        "grid",
+        "grid_that_straddles_dateline_with_shifted_global_etopo_data",
+        "grid_that_straddles_dateline_with_global_srtm15_data",
+    ],
+)
+def test_roundtrip_netcdf(grid_fixture, tmp_path, request):
     """Test that creating a grid, saving it to file, and re-opening it is the same as
     just creating it."""
 
-    # Initialize a Grid object using the initializer
-    grid_init = Grid(
-        nx=10,
-        ny=15,
-        size_x=100.0,
-        size_y=150.0,
-        center_lon=0.0,
-        center_lat=0.0,
-        rot=0.0,
-        topography_source={"name": "ETOPO5"},
-        hmin=5.0,
-    )
+    grid = request.getfixturevalue(grid_fixture)
 
     for file_str in ["test_grid", "test_grid.nc"]:
         # Create a temporary filepath using the tmp_path fixture
@@ -244,7 +342,7 @@ def test_roundtrip_netcdf(tmp_path):
             str(tmp_path / file_str),
         ]:  # test for Path object and str
 
-            grid_init.save(filepath)
+            grid.save(filepath)
 
             filepath = Path(filepath)
 
@@ -252,28 +350,25 @@ def test_roundtrip_netcdf(tmp_path):
             grid_from_file = Grid.from_file(filepath.with_suffix(".nc"))
 
             # Assert that the initial grid and the loaded grid are equivalent (including the 'ds' attribute)
-            assert grid_init == grid_from_file
+            assert grid == grid_from_file
 
             # Clean up the .nc file
             (filepath.with_suffix(".nc")).unlink()
 
 
-def test_roundtrip_yaml(tmp_path):
-    """Test that creating a grid, saving its parameters to yaml file, and re- opening
+@pytest.mark.parametrize(
+    "grid_fixture",
+    [
+        "grid",
+        "grid_that_straddles_dateline_with_shifted_global_etopo_data",
+        "grid_that_straddles_dateline_with_global_srtm15_data",
+    ],
+)
+def test_roundtrip_yaml(grid_fixture, tmp_path, request):
+    """Test that creating a grid, saving its parameters to yaml file, and re-opening
     yaml file creates the same grid."""
 
-    # Initialize a Grid object using the initializer
-    grid_init = Grid(
-        nx=10,
-        ny=15,
-        size_x=100.0,
-        size_y=150.0,
-        center_lon=0.0,
-        center_lat=0.0,
-        rot=0.0,
-        topography_source={"name": "ETOPO5"},
-        hmin=5.0,
-    )
+    grid = request.getfixturevalue(grid_fixture)
 
     # Create a temporary filepath using the tmp_path fixture
     file_str = "test_yaml"
@@ -282,19 +377,29 @@ def test_roundtrip_yaml(tmp_path):
         str(tmp_path / file_str),
     ]:  # test for Path object and str
 
-        grid_init.to_yaml(filepath)
+        grid.to_yaml(filepath)
 
         grid_from_file = Grid.from_yaml(filepath)
 
         # Assert that the initial grid and the loaded grid are equivalent (including the 'ds' attribute)
-        assert grid_init == grid_from_file
+        assert grid == grid_from_file
 
         filepath = Path(filepath)
         filepath.unlink()
 
 
-def test_roundtrip_from_file_yaml(grid, tmp_path):
+@pytest.mark.parametrize(
+    "grid_fixture",
+    [
+        "grid",
+        "grid_that_straddles_dateline_with_shifted_global_etopo_data",
+        "grid_that_straddles_dateline_with_global_srtm15_data",
+    ],
+)
+def test_roundtrip_from_file_yaml(grid_fixture, tmp_path, request):
     """Test that reading a grid from file and then saving it to yaml works."""
+
+    grid = request.getfixturevalue(grid_fixture)
 
     filepath = Path(tmp_path / "test.nc")
     grid.save(filepath)
@@ -551,3 +656,103 @@ def test_plot():
         grid.plot_vertical_coordinate(eta=-1, xi=0)
     with pytest.raises(ValueError, match="Exactly one of"):
         grid.plot_vertical_coordinate(eta=-1, xi=0, s=-1)
+
+
+# Topography tests
+
+
+def test_enclosed_regions():
+    """Test that there are only two connected regions, one dry and one wet."""
+
+    grid = Grid(
+        nx=100,
+        ny=100,
+        size_x=1800,
+        size_y=2400,
+        center_lon=30,
+        center_lat=61,
+        rot=20,
+    )
+
+    reg, nreg = label(grid.ds.mask_rho)
+    npt.assert_equal(nreg, 2)
+
+
+def test_rmax_criterion():
+    grid = Grid(
+        nx=100,
+        ny=100,
+        size_x=1800,
+        size_y=2400,
+        center_lon=30,
+        center_lat=61,
+        rot=20,
+    )
+    r_eta, r_xi = _compute_rfactor(grid.ds.h)
+    rmax0 = np.max([r_eta.max(), r_xi.max()])
+    npt.assert_array_less(rmax0, 0.2)
+
+
+def test_hmin_criterion():
+    grid = Grid(
+        nx=100,
+        ny=100,
+        size_x=1800,
+        size_y=2400,
+        center_lon=30,
+        center_lat=61,
+        rot=20,
+        hmin=5.0,
+    )
+
+    assert grid.hmin == 5.0
+    assert np.less_equal(grid.hmin, grid.ds.h.min())
+
+    grid.update_topography(hmin=10.0)
+
+    assert grid.hmin == 10.0
+    assert np.less_equal(grid.hmin, grid.ds.h.min())
+
+    # this should not do anything
+    grid.update_topography()
+
+    assert grid.hmin == 10.0
+    assert np.less_equal(grid.hmin, grid.ds.h.min())
+
+
+def test_mask_topography_boundary():
+    """Test that the mask and topography along the grid boundaries (north, south, east,
+    west) are identical to the adjacent inland cells."""
+
+    # Create a grid with some land along the northern boundary
+    grid = Grid(
+        nx=10, ny=10, size_x=1000, size_y=1000, center_lon=-20, center_lat=60, rot=0
+    )
+
+    # Toopography
+    np.testing.assert_array_equal(
+        grid.ds.h.isel(eta_rho=0).data, grid.ds.h.isel(eta_rho=1).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.h.isel(eta_rho=-1).data, grid.ds.h.isel(eta_rho=-2).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.h.isel(xi_rho=0).data, grid.ds.h.isel(xi_rho=1).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.h.isel(xi_rho=-1).data, grid.ds.h.isel(xi_rho=-2).data
+    )
+
+    # Mask
+    np.testing.assert_array_equal(
+        grid.ds.mask_rho.isel(eta_rho=0).data, grid.ds.mask_rho.isel(eta_rho=1).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.mask_rho.isel(eta_rho=-1).data, grid.ds.mask_rho.isel(eta_rho=-2).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.mask_rho.isel(xi_rho=0).data, grid.ds.mask_rho.isel(xi_rho=1).data
+    )
+    np.testing.assert_array_equal(
+        grid.ds.mask_rho.isel(xi_rho=-1).data, grid.ds.mask_rho.isel(xi_rho=-2).data
+    )
